snap: understand directories in layout blacklist
The layout blacklist contains string prefixes that are forbidden in
subsequent layout elements. The prefixes are just paths of each layout
elements. To work correctly a layout element that describes a directory
(vs a file) needs to contain the trailing slash. In other words, this
layout is valid:

 - directory /etc/demo (blacklist contribution: /etc/demo/)
 - file /etc/demo.conf (blacklist contribution: /etc/demo.conf)

Because /etc/demo.conf is not a prefix of /etc/demo/, this should be
allowed by the validator.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>