interfaces/apparmor: use EnsureDirState for setupSnapConfineReexec
This patch changes the implementation of setupSnapConfineReexec to
use EnsureDirState and the newly added helper snapConfineFromCore.

This should fix a problem where a profile already exists on disk but has
stale / broken content (for whatever reason). This can happen, for
example, when the profile is restored from backup (such as what we do in
spread tests).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>