cmd/snap: also include tracking channel in list output.
This adds a "Tracking" column to `snap list`, which lists an
abbreviated form of the tracking channel of the snap.

If there is no tracking channel (because it's a local snap), this will
show '-' instead; I also made 'Developer' show up as '-' in this case;
this'll avoid breaking scripts in future.

The abbreviation consists of ellipsising the branch, and dropping the
"stable" from a track/stable channel.