many: add the snapd-generator (#4797)
* many: add the snapd-generator

This patch removes the snap.mount (or var-lib-snapd-snap.mount) unit
and replaces it with a systemd generator that does creates such unit
dynamically at runtime, on early boot, when the system is running
without mount event sharing on the root filesystem.

This fixes a bug where if packaging ships the snap.mount unit and the
unit is stopped (even if it is not started) it would stop all the snap
mount units on the system, resulting in broken snaps.

Since the mount unit is no longer known to packaging it will not be
automatically stopped/started by generated maintainer scripts and thus
avoid the issue.