tests: add support for phased prepare-restore logic
This patch adds a simple system where prepare/restore code
can be placed in individual modules that are executed with a given
priority and can affect each of the phases of the test.

The phases are:
 - prepare_project (only once)
 - restore_project
 - prepare_project_each (for each task in the whole project)
 - restore_project_each
 - prepare_suite (for each suite)
 - restore_suite
 - prepare_suite_each (for each task in a given suite)
 - restore_suite_each

An example module is added along with echo's in each of the phases.
This patch will help to modularize the prepare/restore logic and
hopefully make it more testable and easy to reason about.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>