cmd/snap-update-ns: convert Secure* family of functions into methods
This patch makes SecureMkdirAll and all the friends into a method on the
new Secure type, namely Secure.MkdirAll. This will allow us to have
stateful secure objects that check for configured trespassing on host
filesystem without the need of ugly globals.

This patch is a long, no-op, that will allow the subsequent patch to
meaningfully improve the system.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>