cmd/snap-update-ns: poke hokes when creating source paths for layouts
This patch adjusts the code for poking writable holes in read-only
file-systems. Previously the hole-poking logic was only enabled in the
source path (the thing that is being exposed somewhere else with a bind
mount). For target path (the thing that receives the share) it was always
disabled.

The rationale for that is that snaps cannot expose things that only exists
from its own point of view and are not universally agreed upon (that is,
that also exist in the same location in the main mount name-space).

For an example, let's assume that a pair of snaps, producer and consumer
are using the content interface to share some data. The producer snap
exposes its $SNAP/inexisting, which doesn't exist in the read-only
file-system, to the consumer snaps' $SNAP/existing, which does exist.

At run-time the mount name-space of the producer snap is not altered.
The mount name-space of the consumer snap would be altered to have
effectively this bind mount:

  mount --bind /snap/producer/123/inexisting /snap/consumer/567/existing

Since /snap/producer/123/inexisting doesn't exist we will attempt to
create it but since it it would have to be created on top of a squashfs
that would fail. We might be tempted to create it it using a writable
mimic but this is where a subtle bug would be introduced.

Because both snaps inhabit separate mount name-spaces the consumer snap
would now see a directory that the producer snap doesn't see and cannot
populate with anything useful.

Therefore the use of writable mimics for source paths was disabled.

For layout this restriction does not apply as a snap can choose to expose
$SNAP/inexisting anywhere where layouts permit and it should be able to
see and populate that directory just fine.

Therefore source path now allows the use of writable mimic for layouts.
The code now also handles source and target more symmetrically, as now
both can return synthetic mount changes. A new pair of unit tests is
added for the case where /rofs/source is missing, one for layouts and
one for content interface.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>