Skip test lp-1721518 for arch, snapd is failing to start after reboot
This test is making snapd fail after refresh and the system gets stuck,
so then the system is not removed and it remains alive until it is
manually deleted from the google console.

This seems to be a actual issue, because snapd fails to start after the
system is rebooted. Here some logs are atached:
https://paste.ubuntu.com/p/tKhkMNDbrc/

To reproduce the issue just run: spread google:arch-
linux-64:tests/regression/lp-1721518

An example about the error in travis: https://travis-
ci.org/snapcore/snapd/builds/373167788