cmd/snap-update-ns,tests: mimic the mode and ownership of directories (#5107)
* cmd/snap-update-ns,tests: mimic the mode and ownership of directories

When a writable mimic is constructed it uses a tmpfs to cover the
original directory. That tmpfs comes with some defaults (user root,
group root, mode 1755). Some software may get upset by wide-open
permissions so we should really mimic the original.

This patch changes planWritableMimic to lstat the original directory and
then conveys the mode, uid and gid to the tmpfs mount system call using
auxiliary argument string. Tmpfs is documented to respect them and
appear as mounted with those parameters.

This patch also extends the layout spread test to verify this.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* interfaces/apparmor: allow dac_override for s-u-n

The apparmor profile for snap-update-ns now allows to override DAC
checks since the writable mimic will now properly mimic the mode and
ownership of directories.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>