testutil: record system call errors / return values
This patch expands the facility for recording mocked system calls to also
record errors or return values. This allows to change code that just suggests a
return value to the reader to one that actively checks it:

Code such as:

```
    c.Assert(s.sys.Calls(), DeepEquals, []string{
	`open "/" O_DIRECTORY 0`,       // -> 3
	`openat 3 "foo" O_DIRECTORY 0`, // -> 4
	`close 4`,
	`close 3`,
    })
```

Becomes:

```
    c.Assert(s.sys.RCalls(), DeepEquals, []testutil.CallResult{
	{`open "/" O_DIRECTORY 0`, 3},
	{`openat 3 "foo" O_DIRECTORY 0`, 4},
	{`close 4`, nil},
	{`close 3`, nil},
    })
```

All the internal unit tests are updated to check return values. The old
facility (sys.Calls()) is still available, as a derivative of the new data.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>