interfaces/apparmor: use strict template on openSUSE tumbleweed
This patch changes the apparmor backend so that on openSUSE tumbleweed,
despite using partial apparmor (i.e. mainline kernel) the strict
apparmor template is used.

The mainline kernel used there (4.16 as of this writing) seems to work
well enough to warrant this experiment. Using this will allow us to
explore the feasibility of growing this mode to all distributions with
"sufficient" apparmor (partial but good enough to enable).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>