snapcraft.yaml: add minimal snapcraft.yaml with custom build
This adds a minimal snapcraft.yaml and a custom build plugin.
The snapcraft.yaml only puts the bits into the snapd snap
for now. This means we can start testing the core18/snapd
interplay.

The plugin will just reuse the deb building (which we have to
maintain anyway). The plugin should probably be cleaned up and
pushed upstream as this seems to be useful for projects that
already have a (complicated) deb based build setup.

One issue with the current snapcraft is that it does not deal with
a source dir called "snap" well. We have code located there but
snapcraft will ignore this dir when pulling code or when setting
up the build tree: https://bugs.launchpad.net/bugs/1772584