tests/lib/prepare: be more selective when copying files for the built image
Commit df8d8b21cb56c2d68a1e6faed6ccaf7da6d68af5 introduced a change which
results in the built image being smaller. The way ubuntu-image works, the amount
of free space in the system partition is kept at constant ratio of ~25% of
partition size, meaning smaller partition yields lower absolute free space. With
the system data smaller, and smaller partition size, the amount of free space is
lower than the size of the data we wanted to copy over when preparing an image.

For comparison, the previously built image was ~740MB, with system partition
occupying ~690MB, with ~168MB of free space. Right now, the image is ~400MB,
with system partition taking ~360MB, with ~80MB of free space. The current
prepare code tried to copy ~130MB of data from /home/gopath.

The patch attempts to be more selective about what we copy over to the image,
trying to copy only the necessary bits.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>