data/systemd: add snapd.apparmor.service
This patch adds a new systemd service unit that loads apparmor profiles
on early machine boot-up. It is needed because there are no standardized
apparmor userspace tools that load profiles from
/var/lib/snapd/apparmor/profiles. On Ubuntu and Debian the init scripts
are patched to support this internally but, e.g. on openSUSE they do not.

In general distributions don't ship vanilla upstream init scripts and
until a future release that unifies distributions and makes the list of
profile-holding directories configurable we must do our own thing.

This was requested by jdstrand as an alternative to storing our profiles
in /etc/apparmor.d/.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>