interfaces/udev: call 'udevadm settle --timeout=10' after triggering events
After an interface connect/disconnect, the udev backend calls udevadm to reload
the udev rules and then calls udevadm to trigger events so that existing
devices are udev tagged right away. udevadm trigger will not block so it is
possible that a snap will try to access a device before it is tagged. By
calling 'udevadm settle' we can block for a little while to allow allow the
events we triggered to be handled. Note that udevadm default timeout is 120
seconds, but udevadm settle will return very quickly when only a few events
need to be handled. Arbitrarily set the settle timeout to 10 seconds to not
block the snapd connect/disconnect for too long but to give udev a chance to
process snapd's newly added/removed rules.

Normally you don't need or want to call 'udevadm settle' but in the case of
snapd it is important when considering that AppArmor rules are updated and made
live to the system before the udev backend and before the triggered events are
all handled (since the AppArmor policy may allow more than the device cgroup
would otherwise allow). Ideally we would make these udevadm calls only once
(perhaps with a higher timeout since there could be more events to handle) and
before the AppArmor backend loads its new policy.