cmd/snap-update-ns: discard the concept of segments
This patch discards the concept of segments from the code. Segments were
used as a way to pass around paths exploded to subsequent components but
there were several issues with that. It was non-obvious how it worked,
it was impossible to express relative paths or the root directory and
the code was error-prone with off-by-one errors.

This patch removes all of that with the shift to path iterator for
convenient traversal where necessary as well as the use of
filepath.Split to get the base and the leaf of a given path.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>