tests/lib/prepare-restore: fix upgrade/reboot handling on arch
When running tests on arch we need to perform system upgrade. Since we do now
know what which packages were upgraded, a reboot is issued. Occasionally, when
the base image is recent enough, there may be no updates, in which case there
will be no reboot. Fix the reboot check, but do check if we are running the
installed kernel. Otherwise the tests will fail when attempting to mount snap
images.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>