snapstate,ifstate: wait for pending restarts before auto-connecting
While most of the removal of core-phase-2 handling in
https://github.com/snapcore/snapd/pull/5301 was fine it turns out
we still want to ensure to wait for pending restarts before we connect
interfaces.

This is important because e.g. the new core may have new interfaces
that auto-connect so after core/snapd got installed and requested
a restart in link-snap then the subsequent auto-connect task should
wait until the restart actually happend to ensure that the task
runs with the new auto-connect interfaces.