store: keep all files with link-count > 1 in the cache
The current store download cache cleanup implementation is doing
the cleanup by a simply looking at the file date. We can do better
because all files in the download cache are hardlinked. This means
that when the link count is greater than 1 we get the copy in the
cache for free so there is no need to delete the file.

This is useful e.g. when doing parallel installs because any already
installed snap will be part of the download cache (because the
cache file is just a hardlink to the snap in /var/lib/snapd/snaps).