dirs: improve distro detection for Antegros
Previous attempt to use ID_LIKE=archlinux for detecting Arch Linux like distros
failed. It seems that there is no consensus about using ID/ID_LIKE.

For reference, this is what is contained in /etc/os-release for Arch like
distros:

- Arch:
    ID=arch
    ID_LIKE=archlinux

- Manjaro (no ID_LIKE):
    ID=manjaro

- Antegros:
    ID="antergos"
    ID_LIKE="arch"

Relying on ID_LIKE seems to have little sense since there seems to be no
agreement as to what that variable should be set to.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>