overlord/ifacestate: re-map connection state on load/store from state
This patch changes getConns and setConns to re-map connection state just
after loading from the state and just before being saved to the state.
This is coupled with explicit tests for {get,set}Conns that we apparently
didn't have.

Note that at this stage the code is still dormant since re-mapping is
a no-op until the "snapd" snap becomes the host of the implicit slots.

This is the "back" side of the re-mapping. The "front" side applies to
re-mapping of the API layer. It is provided in the follow-up patch.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>