overlord,daemon: reduce InteraceMapper to SnapMapper
With the insight from the previous patch I have now simplified all of
the interface mapper to just a SnapMapper wich can be thought as
server-side nickname system.

This mostly cuts the amount of boilerplate code that doesn't do anything
we currently need by limiting the re-mapping to just snap names and to
not be able to differentiate plugs and slots anymore.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>