cmd/snap: print unset license as "unset", instead of "unknown
Seems I forgot to circle back on this once the store-side changes were
done.  This wording (unset vs unknown) was agreed to [ages ago][1] and
implemented for the store side but not for local snaps (resulting in a
snap with no license going from "unset" to "unknown" once installed).

More info [in the forum][1].

[1]: https://forum.snapcraft.io/t/3671/26?u=chipaca