daemon, overlord/state: warnings pipeline
This is the backend work for the warnings pipeline. It has some
(possibly ephemeral) debug verbs added as well, to let you play with
it via http if that's your thing.

Missing is:

the frontend work: namely 'snap warnings', 'snap okay', and the
messages after any command.

deleting warnings when they expire.

---

This is actually the third iteration, but the overlord/state code
remains unchanged: the change has been to simplify the way the
warnings summary is returned, thanks to pedronis's insightful insight.