cmd/snap: support `--last=<type>?` to mean "no error on empty"
In some situations a shell script needs to ensure a certain change
can't be initiated, and then check that it hadn't been initiated
before inhibiting. Something like

    # do not auto-refresh for 2 hours
    snap set system refresh.hold=+2h
    # wait for any auto-refreshes already in progress
    snap watch --last=auto-refresh

however if there is no auto-refresh change the above will fail with
`error: no changes of type "auto-refresh" found`.

With this change, simply

    snap watch --last="auto-refresh?"

and the watch will exit with no output and no error.

Note this also works for `abort` and `tasks`.