snap/squashfs: stop printing unsquashfs info to stderr
in #5494 I accidentally included a change that has us printing
unsquashfs info (“Parallel unsquashfs: Using 4 processors” etc.) to
stderr. That change was a debugging aid and not meant for inclusion in
the PR.

This drops the change, and adds a check to the suite teardown to make
sure it doesn't happen again by accident.