store: only files with Nlink > 1 count towards the limit
Before this change we were pruning files with Nlink > 1, but counting
all of them, which could mean we needlessly removed all non-Nlink>1
files in some cases (e.g. the total number of snaps in the system was
bigger than the number of files allowed in the cache). With this
change, only files we would remove are counted as cleanup candidates.