image: ensure "core" is ordered early if base: and core is used
We got a urgent bugreport from a user. The issue is that they are
building a core18 system with a model assertion that has
`required-snaps: [bluez]`. This snap is still the legacy core
snap as its base. With this setup the seed.yaml is written in
a way that puts bluez into the seed before core. This means that
the seeding hangs. This PR adds special handling for "core"
until we fix the ordering of the seed.yaml in a more general
way.

Note that the alternative would be to add "core" to the
required-snaps which is not desirable because bluez will
eventually start using a different base.