interfaces/apparmor: parallel installs support
Introduce support for parallel installation of snaps. Installed snap may have an
optional instance key, in which case it's named <snap>_<instance>. When the
mount namespace of such snap is set up, snap-confine makes adjustments so that
it appears as the snap is running without the instance key. This involves bind
mounting:

* /snap/<snap>_<instance>             -> /snap/<snap>
* /var/<snap>_<instance>              -> /var/snap/<snap>
* /home/<user>/snap/<snap>_<instance> -> /home/<user>/snap/<snap>

The original per snap instance paths are still accesible and advertised through
environment variables SNAP_INSTANCE{,_DATA,_COMMON,_USER_DATA}.

Adjust the deafult template so that the snap can access both the remapped files
and the instance ones.

Note, that XDG_RUNTIME_DIR has been left out from the adjustments as it is not
remapped for now.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>