interfaces: add new "sysfs-name" to i2c interfaces code
We got a request from a customer how needs access to i2c devices
that behave differently from what our i2c code expects. The
device apparently only shows up in /sys/bus/i2c/devices/xxxx
and our code does not support this mode of operation.

This PR provides a way to set a "sysfs-name" that is complementary
to the current "path" we support. If "sysfs-name" is choosen
a apparmor rule is genreated that allows access in
/sys/bus/i2c/devices/$name/**

The "sysfs-name" and "path" cannot be used at the same time in
the same interfce slot.