overlord/snapstate: always clean SnapState when doing Get()
SnapState is (un-)marshalled from/to JSON. Some fields some with omitempty tag,
meaning those will not be included in the output when marshalling and the will
not be cleared in the destination structure when unmarshalling.

Sometimes in the code we reuse snapst for subsequent snapstate.Get() calls.
Should the destination struct not be cleared between the calls, we may end up
with data from the previous calls inside.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>