dirs: fix SnapMountDir inside a Fedora base snap
This patch fixes the algorithm that dertermines the mount point for snaps.  The
old code used to work because it was looking at /etc/os-release and looking for
specific ID or ID_LIKE values. With the introduction of base snaps that can
have arbitrary os-release files we cannot rely on this information alone.

When a process is inside a snap mount namespace that has a base snap mounted as
the root filesystem then the os-release level information is not useful and
must not be used anymore. Inside that environment the snaps are always mounted
at /snap.

On the outside of a mount namespace the current algorithm is correct, it will
identify the distribution and apply any differences mandated by the
distribution policy.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>