cmd/snap: fix some corner-case test setup weirdness
In some cases, when running the tests in a particular way, some bits
of the environment were not set up properly and tests would fail. In
particular colour support was getting isatty for stdout wrong, as it
was never mocked outside of the colour tests. And dbus tests were
failing to unset the session bus env var. It's a mystery how these
things sometimes worked :-) but this should fix it for good.

I've also taken the opportunity to unify our isatty handling (at least
within cmd/snap; there's still code in progress/ doing it separately).