snapcraft: set version information for the snapd snap
The current version-script for the snapcraft.yaml is a bit too
simplistic. It will only look at the debian/changelog. However
for edge versions we do not update debian/changelog so the version
is misleading (e.g. 2.34.3 for current edge).

Instead it should use a combination of the version in the debian
changelog and the git describe output. This means we get:
```
master:       2.35~pre1+721-g385efa21c
release/2.34: 2.34.3+5-gde24f94dd
2.34.3:       2.34.3
```
for the snapd version.