snap: fix mocking of systemkey in snap-run tests
The fix 65ab748 broke the snap run related tests when there is
no snapd running. The reason is that the order of the system-key
mocking and writing of the system key was wrong. Because there
was a cleanup missing this never was a problem. However with the
fixed cleanup it became one. This PR fixes the order and snapd
will build again. This will give us a working "edge" snap again.