overlord/devicestate: DTRT w/a snap proxy to reach a serial vault (#5617)
* overlord/devicestate: DTRT w/a snap proxy to reach a serial vault

Without this change, if `proxy.store` was set on core, and
`device-service.url` was set on the gadget, snapd would attempt to
talk directly to the device service.

This change adds the logic so that, in this case, we talk to the
appropriate endpoint of the proxy, specifying the serial vault we want
in a header.

* overlord/devicestate: use the appropriate endpoint for the proxy

I had mistakenly made snapd talk to the proxy endpoint at …/serial
(the path ending for direct communication with a serial vault) instead
of …/devices when both proxy and serial vault were set.

* overlord/devicestate: only do the proxy+serial dance if the proxy is new enough

Without this we'd always ask the proxy for the serial if it and the
vault were set, and older proxies won't understand the request and
hand out a generic serial instead. With this change, if the prox and
the serial vault are both set, we check the proxy store's version and
only do the full proxy+serial dance if it's new enough.

* overlord/devicestate: add logging to newEnoughProxy

* overlord/devicestate: Split a test, add a comment.

This addresses review feedback from @mvo5 and @pedronis (thanks!).