snap,snap-exec: support command-chain for hooks
When a snap is created by the snapcraft CLI, each hook potentially has
an associated chunk of environment variables defined as well as scripts
that may be run before the hook itself. Currently, this is done by
generating a shell script wrapper for each hook, but this is opaque and
confusing.

Snapd supports an `environment` keyword for apps and hooks as well as a
`command-chain` for apps. Add the last item needed before the shell
wrappers can go away completely: `command-chain` support in hooks.

See forum topic https://forum.snapcraft.io/t/6112 for more details.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>