debian: add missing breaks on comisc
On 18.04 the released version of apt (1.6.1) has a bug that causes
problem on "apt purge snapd". To ensure this won't happen add the
right "Breaks: apt (<< 1.6.3)" to debian/control.

However on cosmic there is the same bug in apt just with a different
version number (<< 1.7.0~alpha2). It looks like the autopkgtest
chroots have an outdated apt on cosmic so we need this breaks on
cosmic as well.