overlord/snapstate: improve feature flag validation
When disabling a feature flag, the common action to do is call:
  snap set system feature.flag=

This has a side effect of the feature flag being set to "" (not false, or nil).
The corecfg validation code accounts for this and type switches correctly. The
snapstate code does not, thus when one sets `experimental.parallel-instances=`
an attempt to install a snap yields this error:

  error: cannot install "test-snapd-tools_foo": internal error: cannot unmarshal snap "core" option
         "experimental.parallel-instances" into *bool: json: cannot unmarshal string into Go value of
         type bool, json: ""

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>