interfaces: rewrite adb-support interface
This patch makes several changes to the adb-support interface. The
permanent slot snippet now grants udev permissions that change the mode
of all the interesting devices to 0666. The connected plug grants udev
and apparmor snippets for tagging the very same devices and for
accessing the relevant parts of /dev and /run/udev/data.

With those changes the slot can be provided by interested applications
directly, without involving the core. This is reflected in the implicit
flags and in the expected policy.

Lastly, since the only phone I could test this against was a BQ Ubuntu
Aquaris E4.5 I added the vendor ID to the list of handled devices.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>