overlord/snapstate: handle common snap dirs when installing/removing snaps with instance key
When installing snaps with instance key set, make sure to create common snap
directories that will be used as mountpoints when running the snap.
Specifically, when installing snap foo_bar, the following directories will be
created:

 - /snap/foo  (bind mount of /snap/foo_bar)
 - /var/snap/foo (bind mount of /var/snap/foo_bar)

Make sure that the directories are cleaned up when the snaps are removed and
there are no more snaps that may require them.

In order to keep the state lockign to minimum, the changeset extends the
snapstate backend operation with 2 additional methods for removing the base snap
directory (/snap/foo) and base snap data directory (/var/snap/foo).

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>