cmd, cmd/snap: better support for non-linux
On non-linux platforms we're never going to be able to re-exec, and we
are not currently going to be able to talk to snapd. This change drops
the nasty warning resulting from the first, and the delay and nasty
error resulting from the second. The result is that instead of

    $ snap list
    2018/08/29 10:51:45.221016 cmd.go:156: cannot read /proc/self/exe: readlink /proc/self/exe: no such file or directory
    [ ... 5 seconds wait ... ]
    error: cannot list snaps: cannot communicate with server: Get http://localhost/v2/snaps: dial unix /run/snapd.socket: connect: no such file or directory

you get

    $ snap list
    Interacting with snapd is not yet supported on CP/M-86.
    This command has been left available for documentation purposes only.