wrappers: remove Wants=network-online.target
When generating units we always and unconditionally pull in
Wants=network-online.target currently. However as reported in

    https://forum.snapcraft.io/t/6063

this will slow down service startup by 2 minutes if any of the
network interfaces is not ready. Instead this PR changes it to
the network.target which is less strict and only means that the
systems network stack is up (and conversely that services should
be stopped before the network is stopped).

Note that this change exiting behaviour and might break existing
snaps. However arguable those are already broken because the
network may go up and down so network-online is not a good
gurantee for something that the software itself should deal
with.