cmd/snap-update-ns: introduce trespassing state tracking
This patch adds a pair of types that are used together as a remedy for
trespassing bug where snap-update-ns writes to places outside of
$SNAP_DATA or $SNAP_COMMON. The types are used as follows:

Assumptions track the global state of the filesystem: directories that
have no write restrictions (like /tmp or $SNAP_DATA) and past mount
operations so that we can detect if a tmpfs was mounted by snapd or not.

Restrictions track state of ongoing operation that may write outside of
$SNAP_DATA or other allowed place. If restriction checking fails the
full path kept by the restrictions object is used to construct a
TrespassingError.

The patch also moves the existing trespassing helpers (isReadOnly and
IsSnapdCreatedPrivateTmpfs) to the same file as the rest of trespassing
code and make both private.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>