daemon, overlord/state: cleanup and rename/refactor pass
Thanks to @niemeyer and @pedronis for their feedback.

This changes AddWarning to Warnf (and has it take args like Sprintf),
changes WarningsToShow to PendingWarnings (including as an action on
the daemon), renames Expired to ExpiredBefore and IsShowable to
ShowAfter, and lastly drops DeleteExpired and adds expired warning
pruning to flatten/unflatten and State.Prune itself.