snap: only show "next" refresh time if its after the hold time
We got a bugreport that the the output of `snap refresh --time`
is confusing when the "core.refresh.hold" option is used. The
issue here is that "next" may display a time before the hold
time set by the user. The reason is that the internal next
will always be calculated and then when its about to run we
check against the hold time. But this is something internal
so an easy fix to avoid confusion is to simply hide "next"
if its before "hold".

See https://forum.snapcraft.io/t/7230