snap-env-generator: do nothing when PATH is unset
When the PATH is not set when the generator runs, simply do nothing
for now. If we print PATH=/snap/bin thats the wrong path and
systemd needs to be smarter to collect paths first. If we add a
default PATH the problem is that this is very inconcistent, see
LP:#1792004.

See also https://bugs.launchpad.net/cloud-images/+bug/1791691