daemon: make error responders not printf when called with 1 argument
Before this change, doing e.g. `InternalError(err.Error())` was a bug:
the error responder would call sprintf on the args, meaning that any
percent sign in the error message would explode into an angry (and
confusing) MISSING in the error message.

In the past I'd caught instances of this bug and fixed the caller.  In
retrospect this was a bad idea, as we now again have instances of the
bug. This change fixes it in the responder: if called with a single
argument, don't call Sprintf. Tadaaaa. No more bug.