interface: add new dotfiles interface
This is a new interface called "dotfiles" (name is a strawman). It
allows to get access to specific files or dirs in the users home
directory. An application might for example use:
```
name: fluxctl
version: 1.0
plugs:
 dotfiles:
  files [.kube/config]
apps:
 app:
  command: fluxctl
  plugs: [dotfiles]
```
Which will allow the fluxctl snap to control the .kube config.

This was requested in:
https://forum.snapcraft.io/t/6948
https://forum.snapcraft.io/t/7062

This gives read/write access right now. This could be split into
read/write if desired. E.g. `read-files:  [.kube/config]`.