interfaces/apparmor: handle overlayfs snippet for snap-update-ns
When running from a live CD which is using overlayfs on the root
filesystem we would get denials from snap-update-ns that tries to
construct content interface connections as appropriate bind mounts.

This is because additional permissions required to support overlayfs
were added to individual application profiles but not to the per-snap
snap-update-ns profile. This is now fixed.

Fixes: https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1794953
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>