data: run snapd.autoimport.service only after seeding
The snapd.autoimport.service will try to import assertions from
removable device that are cold-plugged into the device.

On first-boot there is a race here, if the autoimport service is
run before some key assertions are imported during the seeding
the autoimport (cold-plug) will fail.

This PR fixes this by adding an explicit
 `After=snapd.seeded.service`