spread-shellcheck: use threads to parallelise
Current spread-shellcheck does things serially. The task is of a class
called "embarrassingly parallel", so I did the obvious thing:

Before:

    $ time ./spread-shellcheck spread.yaml tests

    real    1m27.587s
    user    0m53.500s
    sys     0m15.368s

After:

    $ time ./spread-shellcheck spread.yaml tests

    real    0m13.977s
    user    0m56.736s
    sys     0m11.536s

One thing to improve would be to add a lock and hold it around any
call to logging (especially as we do more than one call to logging in
a row and expect them to be adjacent in the output). In practice, not
really a problem right now.