snap: validate plug and slot names
This patch plugs a hole in snap.yaml validation code. We used to check
the plug and slot names because we used to call repo.AddPlug, and
AddSlot but now we call bulk AddSnap (which is faster and atomic) and it
relies on snap.yaml validation entirely, not doing any more checks. The
problem is that we never validated plug and slot names at that level.

This patch fixes that. From now on invalid plug and slot names will be
rejected early on in the validator, preventing usage of such snaps on
the system.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>