osutils: unit tests speedup; introduce «run-checks --short-unit».
In trying to speed up the `osutils` unit tests I found that a
particular test was problematic, as there wasn't a good way of making
it consistently quicker (on my machine it takes between .2s and 40s).

So instead I skip that test if `testing.Short()`, and added a
`--short-unit` option to `run-checks` that runs `go test -short`.
While I was at it I made the short unit test not run with `-coverage`,
which also speeds it up. And of course, the travis "quick" unit test
run is now a `--short-unit`.

So now,

    $ go test -count 10 ./osutil
    ok  	github.com/snapcore/snapd/osutil	118.122s
    $ go test -short -count 10 ./osutil
    ok  	github.com/snapcore/snapd/osutil	5.910s

(I did also shave ~.1s off one other test, FWIW, but it's not one that
runs in travis anyway).